requirements.txt
slack-bolt==1.18.0
slack-sdk==3.27.1
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1

